Refactor TileNavigation component into modular structure
- Split large component into separate files for better organization
- Extract utility functions like document ID parsing and fetching
- Create separate components for NavigationTile and TileCoverImage
- Maintain existing functionality and component behavior
